+++ packages/widget_toolkit/CHANGELOG.md
@@ -1,3 +1,10 @@
+## [0.0.1-dev7]
+- Documentation improvements
+### Breaking Changes:
+- Removed params: `showChangeLanguageBottomSheet(iconRight, errorPanelIcon)`
+- Added parameter `showChangeLanguageBottomSheet(Widget Function(SelectedLanguageModel model, bool loading, BuildContext context)? itemBuilder, required LanguageService service, required Function(LanguageModel language) onChanged, String Function(LanguageModel) translate)`
+  -  [Language Picker API Improvements ticket](https://github.com/Prime-Holding/widget_toolkit/issues/10)
+
 ## [0.0.1-dev6]
 - Documentation improvements
 - `EditAddressWidget` example is added to the example project